pyproject.toml
[project]
name = "esme-base-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "groq>=1.0.0",
    "langfuse>=3.14.1",
    "openinference-instrumentation-smolagents>=0.1.22",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "python-dotenv>=1.2.1",
    "smolagents[litellm,telemetry,toolkit]>=1.24.0",
]
